test: 테스트 코드에 대한 전반적인 수정과 추가 테스트 코드 개발

+++ src/main/java/springJpaBoard/Board/api/apicontroller/MemberApiController.java
@@ -171,10 +171,13 @@ public class MemberApiController {
 
     /* 회원 삭제 */
     @DeleteMapping("/delete/{memberId}")
-    public ResponseEntity deleteMember(@PathVariable Long memberId) {
-        memberService.delete(memberId);
-
-        return ResponseEntity.status(HttpStatus.OK).body("회원 삭제 성공");
+    public ResponseEntity deleteMember(@PathVariable Long memberId, @Login Member loginMember) {
+        Member member = memberService.findOne(memberId);
+        if (memberService.loginValidation(member, loginMember)) {
+            memberService.delete(memberId);
+            return ResponseEntity.status(HttpStatus.OK).body("회원 삭제 성공");
+        }
+        throw new UserException("회원 정보 불일치");
     }
 
     /**
